refactor(trip-mon...selector): Extract component.
